Feature/When To CheckPiece

Approved-by: Przemysław Wiczołek

+++ Player/Models/Player.cs
@@ -539,7 +539,7 @@ namespace Player.Models
             ShamPieceProbability = p.ShamPieceProbability;
             WaitingPlayers = new LinkedList<int>();
             int pomResult = ((BoardSize.y + BoardSize.x) * PenaltiesTimes.Move) + PenaltiesTimes.Pickup + PenaltiesTimes.PutPiece;
-            if (((BoardSize.x * GoalAreaSize) / NumberOfPlayers.Allies) * (pomResult + PenaltiesTimes.Response + PenaltiesTimes.Ask) < ((BoardSize.x / NumberOfPlayers.Allies) * GoalAreaSize * pomResult)
+            if (((double)(BoardSize.x * GoalAreaSize) / NumberOfPlayers.Allies) * (pomResult + PenaltiesTimes.Response + PenaltiesTimes.Ask) < (((double)BoardSize.x / NumberOfPlayers.Allies) * GoalAreaSize * pomResult)
             && pomResult < (NumberOfPlayers.Allies * PenaltiesTimes.Response) && !(2 * NumberOfPlayers.Allies < BoardSize.x))
             {
                 IsCommunicatinonWorthy = true;
